chore; separate plan and add on pricing content
