+++ translations/base.json
@@ -4196,7 +4196,7 @@
   "text_1779972384579cgilv8fpjw6": "Failed to create quote",
   "text_1779802343219a1cl5ckvtrn": "Pricing",
   "text_1779802343219rul1jvs7170": "Insert a pricing block",
-  "text_1779802343220xh5jm32or13": "Add-ons",
+  "text_1779802343220xh5jm32or13": "Add an add-on",
   "text_17798023432203q6hytdp7om": "Select add-ons",
   "text_1779805897126caxqtv14ctd": "Add pricing",
   "text_17799586575620rdqef1d7dq": "Pricing for one-off invoice",
